requirements.txt
mysqlclient
python-dotenv
django
django-rest-auth
djangorestframework
django-enumfields
drf-enum-field
requests